feat: implement createApp function and refactor OpenAPI document generation
