+++ api/scripts/build-openapi.ts
@@ -2,18 +2,23 @@ import { existsSync } from "node:fs";
 import { mkdir, writeFile } from "node:fs/promises";
 import path from "node:path";
 import process from "node:process";
-import { getOpenAPIDocument } from "../src";
+import { createApp } from "../src/app";
 import { buildOpenApiConfig } from "../src/openapi";
 
 const root = path.resolve(import.meta.dirname, "../");
 
 async function run() {
-  const obj = getOpenAPIDocument(buildOpenApiConfig("x.y.z", [
+  // Create the app with all routes using the shared function
+  const app = createApp();
+
+  const config = buildOpenApiConfig("x.y.z", [
     {
       url: "https://unicode-api.luxass.dev",
       description: "Production Environment",
     },
-  ]));
+  ]);
+
+  const obj = app.getOpenAPIDocument(config);
 
   if (!existsSync(path.join(root.toString(), "./node_modules/.openapi"))) {
     await mkdir(path.join(root.toString(), "./node_modules/.openapi"), { recursive: true });
